Added add'l logic to setting up symlinks
Now we check to see if the dotfile exists, if it does, then we check if
it is already set as the proper symlink to our dotfiles. If it is not
properly set, we then backup (mv) the existing file or directory to
.dotfiles_old, then we create the dotfile symlink as it should be.

Using this new method keeps from filling up our .dotfiles_old directory
with a bunch of backed up dotfiles which are nothing more than symlinks
to our desired .dotfiles location. The previous method was not efficient
and purely just did a backup (mv) no matter what and then recreated the
symlink.

+++ setup.sh
@@ -24,23 +24,31 @@ cd $DOTFILES_DIR || exit
 for dotfile in "${dotfiles[@]}"
 do
   if [ -f $HOME/$dotfile ]; then
-    echo "Backing up $dotfile to $DOTFILES_DIR_BACKUP/$dotfile.$timestamp..."
-    mv -f $HOME/$dotfile $DOTFILES_DIR_BACKUP/$dotfile.$timestamp
-    echo "Done"
+    if [[ ! "$HOME/$dotfile" -ef "$DOTFILES_DIR/$dotfile" ]]; then
+      echo "Backing up $dotfile to $DOTFILES_DIR_BACKUP/$dotfile.$timestamp..."
+      mv -f $HOME/$dotfile $DOTFILES_DIR_BACKUP/$dotfile.$timestamp
+      echo "Done"
+    fi
   fi
   if [ -d $HOME/$dotfile ]; then
-    echo "Backing up $dotfile directory to $DOTFILES_DIR_BACKUP/$dotfile.$timestamp..."
-    mv -f $HOME/$dotfile $DOTFILES_DIR_BACKUP/$dotfile.$timestamp
-    echo "Done"
+    if [[ ! "$HOME/$dotfile" -ef "$DOTFILES_DIR/$dotfile" ]]; then
+      echo "Backing up $dotfile directory to $DOTFILES_DIR_BACKUP/$dotfile.$timestamp..."
+      mv -f $HOME/$dotfile $DOTFILES_DIR_BACKUP/$dotfile.$timestamp
+      echo "Done"
+    fi
   fi
 done
 
 cd $HOME || exit
 for dotfile in "${dotfiles[@]}"
 do
-  echo "Creating symlink for $DOTFILES_DIR/$dotfile..."
-  ln -s $DOTFILES_DIR/$dotfile .
-  echo "Done"
+  if [ ! -e $dotfile ]; then
+    echo "Creating symlink for $DOTFILES_DIR/$dotfile..."
+    ln -s $DOTFILES_DIR/$dotfile .
+    echo "Done"
+  else
+    echo "$HOME/$dotfile symlink already exists..."
+  fi
 done
 
 # Check if bash-git-prompt is already installed and install if not
